feat(celery beat): setup celery beat service for periodic tasks
https://app.asana.com/0/308727253154577/321388985457824
