+++ .ebextensions.worker/worker.config
@@ -3,19 +3,15 @@ commands:
     command: "/opt/python/run/venv/bin/pip install --upgrade pip"
 
 option_settings:
-  aws:elasticbeanstalk:application:environment:
-    DJANGO_SETTINGS_MODULE: settings.staging
 
 container_commands:
+  celeryd:
+    command: "sudo service celeryd restart"
+  celerybeat:
+    command: "sudo service celerybeat restart"
 
 services:
   sysvinit:
     rabbitmq-server:
       enabled: "true"
       ensureRunning: "true"
-    celeryd:
-      enabled: "true"
-      ensureRunning: "true"
-      sources:
-        - "selstagram2"
-        - "instagram"
