+++ celery_daemon/etc/default/celerybeat
@@ -1,0 +1,17 @@
+#!/usr/bin/env bash
+
+# Absolute or relative path to the 'celery' command:
+CELERY_BIN="celery"
+#CELERY_BIN="/virtualenvs/def/bin/celery"
+
+# App instance to use
+# comment out this line if you don't use an app
+CELERY_APP="selstagram2.celerybeat"
+# or fully qualified:
+#CELERY_APP="proj.tasks:app"
+
+# Where to chdir at start.
+CELERYBEAT_CHDIR="/opt/python/current/app"
+
+# Extra arguments to celerybeat
+CELERYBEAT_OPTS="--schedule=/var/run/celery/celerybeat-schedule --broker=amqp://guest:guest@localhost:5672//"
