+++ selstagram2/instagram/tasks.py
@@ -2,8 +2,6 @@
 # -*- coding: utf-8 -*-
 from urllib import request
 
-from celery.schedules import crontab
-
 from selstagram2.celery import app
 
 
@@ -13,21 +11,6 @@ def download(url):
     print(file_path, http_message)
 
     return file_path
-
-
-@app.on_after_configure.connect
-def setup_periodic_tasks(sender, **kwargs):
-    # Calls test('hello') every 10 seconds.
-    sender.add_periodic_task(10.0, test.s('hello'), name='add every 10')
-
-    # Calls test('world') every 30 seconds
-    sender.add_periodic_task(30.0, test.s('world'), expires=10)
-
-    # Executes every Monday morning at 7:30 a.m.
-    sender.add_periodic_task(
-        crontab(hour=7, minute=30, day_of_week=1),
-        test.s('Happy Mondays!'),
-    )
 
 
 @app.task
